Update README.md
It was not clear that the directory folder required was not already somewhere on my C: drive included as part of the Power BI standard install. After searching for its location and then googling I found that it was a folder which the user needs to create, hence I believe the install instructions needed re-wording.